requirements.txt
opencv-python==4.7.0.68
mediapipe==0.10.21
scikit-learn>=1.4.0
streamlit
Pillow
